backend/requirements.txt
django==3.2.16
djangorestframework==3.14.0
django-colorfield==0.7.2
djoser==2.1.0
django-cors-headers==3.13.0
django-extra-fields==3.0.2
django-filter==22.1
psycopg2==2.9.5